Created cli_test.py in tests, renamed former cli.py and cli_test.py in krux_cloud_health
